Readonly updates to the readonly branch

+++ src/modules/Elsa.Workflows.Api/Endpoints/WorkflowDefinitions/Delete/Endpoint.cs
@@ -1,27 +1,55 @@
 using Elsa.Abstractions;
+using Elsa.Common.Models;
 using Elsa.Workflows.Api.Constants;
 using Elsa.Workflows.Management.Contracts;
+using Elsa.Workflows.Management.Filters;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Elsa.Workflows.Api.Endpoints.WorkflowDefinitions.Delete;
 
 internal class Delete : ElsaEndpoint<Request>
 {
     private readonly IWorkflowDefinitionManager _workflowDefinitionManager;
+    private readonly IAuthorizationService _authorizationService;
+    private readonly IWorkflowDefinitionStore _store;
 
-    public Delete(IWorkflowDefinitionManager workflowDefinitionManager)
+    public Delete(IWorkflowDefinitionManager workflowDefinitionManager, IAuthorizationService authorizationService, IWorkflowDefinitionStore store)
     {
         _workflowDefinitionManager = workflowDefinitionManager;
+        _authorizationService = authorizationService;
+        _store = store;
     }
 
     public override void Configure()
     {
         Delete("/workflow-definitions/{definitionId}");
         ConfigurePermissions("delete:workflow-definitions");
-        Policies(AuthorizationPolicies.ReadOnlyPolicy);
     }
 
     public override async Task HandleAsync(Request request, CancellationToken cancellationToken)
     {
+        var filter = new WorkflowDefinitionFilter
+        {
+            DefinitionId = request.DefinitionId,
+            VersionOptions = VersionOptions.Latest
+        };
+
+        var definition = await _store.FindAsync(filter, cancellationToken);
+
+        if (definition == null)
+        {
+            await SendNotFoundAsync(cancellationToken);
+            return;
+        }
+
+        var authorizationResult = _authorizationService.AuthorizeAsync(User, definition, AuthorizationPolicies.NotReadOnlyPolicy);
+
+        if (!authorizationResult.Result.Succeeded)
+        {
+            await SendForbiddenAsync(cancellationToken);
+            return;
+        }
+
         var result = await _workflowDefinitionManager.DeleteByDefinitionIdAsync(request.DefinitionId, cancellationToken);
 
         if (result == 0)
